Move code to RiaEnsembleImportTools
