+++ ApplicationLibCode/Application/Tools/Ensemble/RiaEnsembleImportTools.h
@@ -1,0 +1,37 @@
+/////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (C) 2025     Equinor ASA
+//
+//  ResInsight is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  ResInsight is distributed in the hope that it will be useful, but WITHOUT ANY
+//  WARRANTY; without even the implied warranty of MERCHANTABILITY or
+//  FITNESS FOR A PARTICULAR PURPOSE.
+//
+//  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html>
+//  for more details.
+//
+/////////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "Summary/RiaSummaryDefines.h"
+
+#include <QStringList>
+
+class RimSummaryCase;
+
+namespace RiaEnsembleImportTools
+{
+struct CreateConfig
+{
+    RiaDefines::FileType fileType;
+    bool                 ensembleOrGroup;
+    bool                 allowDialogs;
+};
+std::pair<bool, std::vector<RimSummaryCase*>> createSummaryCasesFromFiles( const QStringList& fileNames, CreateConfig createConfig );
+
+} // namespace RiaEnsembleImportTools
