+++ ApplicationLibCode/Commands/RicImportSummaryGroupFeature.cpp
@@ -18,6 +18,7 @@
 
 #include "RicImportSummaryGroupFeature.h"
 
+#include "Ensemble/RiaEnsembleImportTools.h"
 #include "RiaGuiApplication.h"
 #include "Summary/RiaSummaryDefines.h"
 
@@ -47,8 +48,8 @@ void RicImportSummaryGroupFeature::onActionTriggered( bool isChecked )
 
     RiaDefines::FileType fileType = RicRecursiveFileSearchDialog::mapSummaryFileType( result.fileType );
 
-    RicImportSummaryCasesFeature::CreateConfig createConfig{ .fileType = fileType, .ensembleOrGroup = true, .allowDialogs = true };
-    auto [isOk, cases] = RicImportSummaryCasesFeature::createSummaryCasesFromFiles( fileNames, createConfig );
+    RiaEnsembleImportTools::CreateConfig createConfig{ .fileType = fileType, .ensembleOrGroup = true, .allowDialogs = true };
+    auto [isOk, cases] = RiaEnsembleImportTools::createSummaryCasesFromFiles( fileNames, createConfig );
 
     RicImportSummaryCasesFeature::addSummaryCases( cases );
     RicImportEnsembleFeature::groupSummaryCases( cases, "", RiaDefines::EnsembleGroupingMode::FMU_FOLDER_STRUCTURE, false );
